backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.50
psycopg2-binary==2.9.7
alembic==1.11.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==1.10.9
python-multipart==0.0.6
email-validator==1.3.1
